requirements.txt
Django==4.2.11
whitenoise==6.6.0
pillow==10.3.0
numpy==1.26.0
matplotlib==3.9.0
opencv-python==4.9.0.80
inference_sdk==0.11.2
google-generativeai==0.5.4
